+++ src/main/java/com/github/kmfisk/workdog/entity/GermanShepherdEntity.java
@@ -29,8 +29,8 @@ public class GermanShepherdEntity extends ProtectionDogEntity {
     }
 
     @Override
-    public double sprintPredisposition() {
-        return 2.0D;
+    public float sprintPredisposition() {
+        return 2.0F;
     }
 
     @Override
