+++ src/main/java/com/github/kmfisk/workdog/entity/JackRussellTerrierEntity.java
@@ -29,8 +29,8 @@ public class JackRussellTerrierEntity extends TerrierDogEntity {
     }
 
     @Override
-    public double sprintPredisposition() {
-        return 2.0D;
+    public float sprintPredisposition() {
+        return 2.0F;
     }
 
     @Override
